requirements.txt
flask>=2.0.3
pillow>=9.0.0
numpy>=1.22.0
numba>=0.55.0
gunicorn>=20.1.0